Moved request info to seperate class/module
