+++ lib/Cache.js
@@ -1,8 +1,6 @@
 var Promise = require("bluebird");
 var Map = require("collections/map");
 
-//////////////////////////////////////////      Cache           //////////////////////////////////////////
-
 var Cache = function () {
     this.requestToResponse = new Map();
 };
@@ -106,18 +104,5 @@ module.exports = Cache;
 
 //////////////////////////////////////////      RequestInfo     //////////////////////////////////////////
 
-var RequestInfo = function (method, url) {
-    this.key = method + '^' + url; // ^ is not valid in method or url
-};
 
-var NOT_CACHEABLE = new RequestInfo(null, null);
-
-RequestInfo.prototype.isCacheable = function () {
-    return this.key !== NOT_CACHEABLE.key;
-};
-
-module.exports = {
-    Cache: Cache,
-    RequestInfo: RequestInfo
-};
