+++ lib/configuration.js
@@ -1,13 +1,14 @@
 var Promise = require('bluebird');
+var websocket = require('websocket-stream');
+var http2 = require('http2');
+var util = require('util');
+var EventEmitter = require('events').EventEmitter;
+
+var RequestInfo = require('./request-info.js');
+var Cache = require('./cache.js');
 var resolvePort = require("./utils.js").resolvePort;
 var parseUrl = require("./utils.js").parseUrl;
 var getOrigin = require("./utils.js").getOrigin;
-var websocket = require('websocket-stream');
-var http2 = require('http2');
-var RequestInfo = require('./cache.js').RequestInfo;
-var Cache = require('./cache.js').Cache;
-var util = require('util');
-var EventEmitter = require('events').EventEmitter;
 var InvalidStateError = require('./utils').InvalidStateError;
 
 /*
@@ -25,11 +26,11 @@ util.inherits(Configuration, EventEmitter);
 
 var confProto = Configuration.prototype;
 
-confProto.increment = function () {
+confProto.configuring = function () {
     this.activeConfigurationCnt++;
 };
 
-confProto.decrement = function () {
+confProto.configured = function () {
     this.activeConfigurationCnt--;
     if (this.activeConfigurationCnt === 0) {
         this.emit('completed');
@@ -40,7 +41,11 @@ confProto.isConfiguring = function () {
     return this.activeConfigurationCnt > 0;
 };
 
-
+/**
+ * Returns the proxy transport for any url
+ * @param url
+ * @returns {*}
+ */
 confProto.getTransport = function (url) {
     if (!this.activeWSConnections[url] || !this.activeWSConnections[url].writable) {
         // TODO, maybe enable perMessageDeflate in production
@@ -50,8 +55,45 @@ confProto.getTransport = function (url) {
     return this.activeWSConnections[url];
 };
 
+/**
+ * Returns the proxy transport url for any origin
+ * @param origin
+ * @returns {*}
+ */
 confProto.getTransportUrl = function(origin){
     return this.proxyMap[origin];
+};
+
+confProto.onPush = function(pushRequest, origin) {
+    var self = this;
+    this.cache.match(pushRequest).then(function (response) {
+        console.warn("Server pushed an already cached result or an un cache-able response: " + response);
+        pushRequest.close();
+    }, function () {
+        var requestInfo = origin ? new RequestInfo(pushRequest.method, origin + pushRequest.url) : new RequestInfo(pushRequest.method, getOrigin(pushRequest) + pushRequest.url);
+        self.cache.put(requestInfo, new Promise(function (resolve, reject) {
+
+            pushRequest.on('response', function (response) {
+
+                response.on('data', function (data) {
+                    if (response.data) {
+                        response.data += data;
+                    } else {
+                        response.data = data;
+                    }
+                });
+
+                response.on('end', function () {
+                    resolve(response);
+                });
+
+                response.on('error', function (e) {
+                    reject(e);
+                });
+
+            });
+        }));
+    });
 };
 
 // open h2 pull channel
@@ -103,7 +145,7 @@ confProto.addConfig = function (config) {
 confProto.fetchConfig = function (url) {
     var xhr = new XMLHttpRequest();
     xhr._open('GET', url);
-    this.increment();
+    this.configuring();
     var self = this;
     xhr.addEventListener("readystatechange", function () {
         if (xhr.readyState === XMLHttpRequest.DONE) {
@@ -112,7 +154,7 @@ confProto.fetchConfig = function (url) {
                 throw new InvalidStateError('Failed to load configuration ' + url + ', status code: ' + status);
             }
             self.addConfig(xhr.response);
-            self.decrement();
+            self.configured();
         }
     }, true);
     xhr._send();
@@ -134,36 +176,5 @@ confProto.getCache = function () {
     return this.cache;
 };
 
-confProto.onPush = function(pushRequest, origin) {
-    var self = this;
-    this.cache.match(pushRequest).then(function (response) {
-        console.warn("Server pushed an already cached result or an un cache-able response: " + response);
-        pushRequest.close();
-    }, function () {
-        var requestInfo = origin ? new RequestInfo(pushRequest.method, origin + pushRequest.url) : new RequestInfo(pushRequest.method, getOrigin(pushRequest) + pushRequest.url);
-        self.cache.put(requestInfo, new Promise(function (resolve, reject) {
-
-            pushRequest.on('response', function (response) {
-
-                response.on('data', function (data) {
-                    if (response.data) {
-                        response.data += data;
-                    } else {
-                        response.data = data;
-                    }
-                });
-
-                response.on('end', function () {
-                    resolve(response);
-                });
-
-                response.on('error', function (e) {
-                    reject(e);
-                });
-
-            });
-        }));
-    });
-};
 
 module.exports = Configuration;
