+++ lib/http2-proxy.js
@@ -1,4 +1,5 @@
 (function () {
+
     var Configuration = require('./configuration.js');
     var enableXHROverH2 = require('./xhr.js').enableXHROverH2;
