+++ lib/xhr.js
@@ -2,7 +2,7 @@ var Promise = require('bluebird');
 var http2 = require('http2');
 var redefine = require('./utils').redefine;
 var dataToType = require('./utils').dataToType;
-var RequestInfo = require('./cache.js').RequestInfo;
+var RequestInfo = require('./request-info.js');
 var parseUrl = require('./utils').parseUrl;
 var getOrigin = require('./utils').getOrigin;
 var InvalidStateError = require('./utils').InvalidStateError;
